+++ demo/margin.js
@@ -31,7 +31,9 @@ const marginMain = async () => {
       };
 
       // Call the borrowMargin function with the requestData object
-      const response = await API.rest.Margin.MarginV3.marginBorrowV3(requestData);
+      const response = await API.rest.Margin.MarginV3.marginBorrowV3(
+        requestData
+      );
 
       // Log the response from the API
       console.log("Borrow Margin Response:", response);
@@ -305,11 +307,12 @@ const marginMain = async () => {
    */
   async function getRedemptionOrdersV3() {
     try {
-      const result = await API.rest.Margin.MarginV3.updatePurchaseOrderInterestRateV3({
-        currency: "BTC",
-        purchaseOrderNo: "5da6dba0f943c0c81f5d5db5",
-        interestRate: "0.05", // Updated interest rate
-      });
+      const result =
+        await API.rest.Margin.MarginV3.updatePurchaseOrderInterestRateV3({
+          currency: "BTC",
+          purchaseOrderNo: "5da6dba0f943c0c81f5d5db5",
+          interestRate: "0.05", // Updated interest rate
+        });
       console.log("purchaseOrderNo number:", result.data.purchaseOrderNo);
     } catch (error) {
       console.error("Error fetching purchaseOrderNo:", error.message);
@@ -317,6 +320,316 @@ const marginMain = async () => {
   }
 
   getRedemptionOrdersV3();
+
+  /**
+ * @name demoGetCrossMarginTradingPairs
+ * @description Demo function to get the configuration of cross margin trading pairs
+ * @updateTime 05/29/24
+ */
+  async function demoGetCrossMarginTradingPairs() {
+    try {
+      const result =
+        await API.rest.Margin.MarginV3.getCrossMarginTradingPairs();
+
+      console.log("Cross Margin Trading Pairs Configuration:", result);
+    } catch (error) {
+      console.error(
+        "Error fetching cross margin trading pairs configuration:",
+        error.message
+      );
+    }
+  }
+
+  demoGetCrossMarginTradingPairs();
+
+  /**
+ * @name demoUpdateLeverageMultiplier
+ * @description Demo function to modify the leverage multiplier
+ * @updateTime 05/29/24
+ */
+  async function demoUpdateLeverageMultiplier() {
+    try {
+      await API.rest.Margin.MarginV3.updateLeverageMultiplier({
+        leverage: "5.00", // New leverage multiplier
+
+        symbol: "BTC-USDT", // Specify for isolated margin
+
+        isIsolated: true, // Set to true for isolated margin
+      });
+
+      console.log("Leverage multiplier updated successfully.");
+    } catch (error) {
+      console.error("Error updating leverage multiplier:", error.message);
+    }
+  }
+
+  demoUpdateLeverageMultiplier();
+
+  /**
+ * @name demoPlaceHfMarginOrder
+ * @description Demo function to place a high-frequency margin trading order
+ * @updateTime 05/29/24
+ */
+  async function demoPlaceHfMarginOrder() {
+    try {
+      const orderDetails = {
+        clientOid: "unique-order-id-12345",
+        side: "buy",
+        symbol: "BTC-USDT",
+        type: "limit",
+        price: "30000",
+        size: "0.01",
+        isIsolated: false,
+        autoBorrow: true,
+        autoRepay: true,
+      };
+
+      const result = await API.rest.Margin.MarginV3.placeHfMarginOrder(
+        orderDetails
+      );
+
+      console.log("Order placed successfully:", result);
+    } catch (error) {
+      console.error("Error placing order:", error.message);
+    }
+  }
+
+  demoPlaceHfMarginOrder();
+
+  /**
+ * @name demoTestHfMarginOrder
+ * @description Demo function to test placing a high-frequency margin trading order
+ * @updateTime 05/29/24
+ */
+  async function demoTestHfMarginOrder() {
+    try {
+      const orderDetails = {
+        clientOid: "unique-order-id-12345",
+        side: "buy",
+        symbol: "BTC-USDT",
+        type: "limit",
+        price: "30000",
+        size: "0.01",
+        isIsolated: false,
+        autoBorrow: true,
+        autoRepay: true,
+      };
+
+      const result = await API.rest.Margin.MarginV3.testHfMarginOrder(
+        orderDetails
+      );
+
+      console.log("Order test successful:", result);
+    } catch (error) {
+      console.error("Error testing order:", error.message);
+    }
+  }
+
+  demoTestHfMarginOrder();
+
+  /**
+ * @name demoCancelHfMarginOrder
+ * @description Demo function to cancel a high-frequency margin trading order by orderId
+ * @updateTime 05/29/24
+ */
+  async function demoCancelHfMarginOrder() {
+    try {
+      const orderId = "5bd6e9286d99522a52e458de";
+      const symbol = "ETH-BTC";
+      const result = await API.rest.Margin.MarginV3.cancelHfMarginOrder(
+        orderId,
+        symbol
+      );
+
+      console.log("Order canceled successfully:", result);
+    } catch (error) {
+      console.error("Error canceling order:", error.message);
+    }
+  }
+
+  demoCancelHfMarginOrder();
+
+  /**
+ * @name demoCancelHfMarginOrderByClientOid
+ * @description Demo function to cancel a high-frequency margin trading order by clientOid
+ * @updateTime 05/29/24
+ */
+  async function demoCancelHfMarginOrderByClientOid() {
+    try {
+      const clientOid = "6d539dc614db3";
+
+      const symbol = "ETH-BTC";
+
+      const result =
+        await API.rest.Margin.MarginV3.cancelHfMarginOrderByClientOid(
+          clientOid,
+          symbol
+        );
+
+      console.log("Order canceled successfully:", result);
+    } catch (error) {
+      console.error("Error canceling order:", error.message);
+    }
+  }
+
+  demoCancelHfMarginOrderByClientOid();
+
+  /**
+ * @name demoCancelAllHfMarginOrdersBySymbol
+ * @description Demo function to cancel all open high-frequency margin trading orders by symbol
+ * @updateTime 05/29/24
+ */
+  async function demoCancelAllHfMarginOrdersBySymbol() {
+    try {
+      const symbol = "ETH-BTC";
+      const tradeType = "MARGIN_TRADE";
+      const result =
+        await API.rest.Margin.MarginV3.cancelAllHfMarginOrdersBySymbol(
+          symbol,
+          tradeType
+        );
+
+      console.log("All orders canceled successfully:", result);
+    } catch (error) {
+      console.error("Error canceling all orders:", error.message);
+    }
+  }
+
+  demoCancelAllHfMarginOrdersBySymbol();
+
+  /**
+ * @name demoGetActiveHfMarginOrders
+ * @description Demo function to get the list of active high-frequency margin trading orders
+ * @updateTime 05/29/24
+ */
+  async function demoGetActiveHfMarginOrders() {
+    try {
+      const tradeType = "MARGIN_TRADE";
+
+      const symbol = "BTC-ETH";
+
+      const result = await API.rest.Margin.MarginV3.getActiveHfMarginOrders(
+        tradeType,
+        symbol
+      );
+
+      console.log("Active HF Margin Orders:", result);
+    } catch (error) {
+      console.error("Error fetching active HF margin orders:", error.message);
+    }
+  }
+
+  demoGetActiveHfMarginOrders();
+
+  /**
+ * @name demoGetFilledHfMarginOrders
+ * @description Demo function to get the list of filled high-frequency margin trading orders
+ * @updateTime 05/29/24
+ */
+  async function demoGetFilledHfMarginOrders() {
+    try {
+      const params = {
+        tradeType: "MARGIN_TRADE",
+        symbol: "BTC-ETH",
+        side: "buy",
+        type: "limit",
+        startAt: Date.now() - 24 * 60 * 60 * 1000, // 24 hours ago
+        endAt: Date.now(),
+      };
+
+      const result = await API.rest.Margin.MarginV3.getFilledHfMarginOrders(
+        params
+      );
+
+      console.log("Filled HF Margin Orders:", result);
+    } catch (error) {
+      console.error("Error fetching filled HF margin orders:", error.message);
+    }
+  }
+
+  demoGetFilledHfMarginOrders();
+
+  /**
+ * @name demoGetHfOrderDetails
+ * @description Demo function to call getHfOrderDetails and log the result
+ * @updateTime 10/05/23
+ */
+  async function demoGetHfOrderDetails() {
+    try {
+      const result = await API.rest.Margin.MarginV3.getHfOrderDetails({
+        orderId: "5c35c02703aa673ceec2a168",
+        symbol: "ETH-BTC",
+      });
+
+      console.log("HF Order Details:", result);
+    } catch (error) {
+      console.error("Error fetching HF order details:", error.message);
+    }
+  }
+
+  demoGetHfOrderDetails();
+
+  /**
+ * @name demoGetHfOrderDetailsByClientOid
+ * @description Demo function to call getHfOrderDetailsByClientOid and log the result
+ * @updateTime 10/05/23
+ */
+  async function demoGetHfOrderDetailsByClientOid() {
+    try {
+      const result = await API.rest.Margin.MarginV3.getHfOrderDetailsByClientOid({
+        clientOid: "6d539dc614db312",
+        symbol: "ETH-BTC",
+      });
+
+      console.log("HF Order Details by Client OID:", result);
+    } catch (error) {
+      console.error(
+        "Error fetching HF order details by clientOid:",
+        error.message
+      );
+    }
+  }
+
+  demoGetHfOrderDetailsByClientOid();
+  /**
+ * @name demoGetHfTransactionRecords
+ * @description Demo function to call getHfTransactionRecords and log the result
+ * @updateTime 10/05/23
+ */
+  async function demoGetHfTransactionRecords() {
+    try {
+      const result = await API.rest.Margin.MarginV3.getHfTransactionRecords({
+        symbol: "BTC-USDT",
+        tradeType: "MARGIN_TRADE",
+        limit: 5,
+      });
+
+      console.log("HF Transaction Records:", result);
+    } catch (error) {
+      console.error("Error fetching HF transaction records:", error.message);
+    }
+  }
+
+  demoGetHfTransactionRecords();
+
+  /**
+ * @name demoGetActiveHfOrderSymbols
+ * @description Demo function to call getActiveHfOrderSymbols and log the result
+ * @updateTime 10/05/23
+ */
+  async function demoGetActiveHfOrderSymbols() {
+    try {
+      const result = await API.rest.Margin.MarginV3.getActiveHfOrderSymbols({
+        tradeType: "MARGIN_TRADE",
+      });
+
+      console.log("Active HF Order Symbols:", result);
+    } catch (error) {
+      console.error("Error fetching active HF order symbols:", error.message);
+    }
+  }
+
+  demoGetActiveHfOrderSymbols();
 };
 
 // run rest marginMain